CCM-8993: use client component
