+++ tests/test-team/pages/letter/template-mgmt-preview-letter-page.ts
@@ -7,27 +7,10 @@ export class TemplateMgmtPreviewLetterPage extends TemplateMgmtPreviewBasePage {
   public static readonly urlRegexp = new RegExp(
     /\/templates\/preview-letter-template\/([\dA-Fa-f-]+)(?:\?from=edit)?$/
   );
-
-  public readonly editRadioOption: Locator;
-
-  public readonly submitRadioOption: Locator;
-
-  public readonly errorSummary: Locator;
-
-  public readonly messageText: Locator;
-
   public readonly continueButton: Locator;
 
   constructor(page: Page) {
     super(page);
-    this.editRadioOption = page.locator(
-      '[id="previewLetterTemplateAction-letter-edit"]'
-    );
-    this.submitRadioOption = page.locator(
-      '[id="previewLetterTemplateAction-letter-submit"]'
-    );
-    this.errorSummary = page.locator('[class="nhsuk-error-summary"]');
-    this.messageText = page.locator('[id="preview-content-message"]');
     this.continueButton = page.locator(
       '[id="preview-letter-template-submit-button"]'
     );
