+++ tests/test-team/template-mgmt-component-tests/letter/template-mgmt-preview-letter-page.component.spec.ts
@@ -64,10 +64,6 @@ test.describe('Preview Letter template Page', () => {
       `${baseURL}/templates/${TemplateMgmtPreviewLetterPage.pageUrlSegment}/${templates.valid.id}`
     );
 
-    await expect(previewLetterTemplatePage.editRadioOption).not.toBeChecked();
-
-    await expect(previewLetterTemplatePage.submitRadioOption).not.toBeChecked();
-
     await expect(previewLetterTemplatePage.pageHeader).toContainText(
       'test-template-letter'
     );
@@ -111,33 +107,6 @@ test.describe('Preview Letter template Page', () => {
       await previewLetterTemplatePage.loadPage('/fake-template-id');
 
       await expect(page).toHaveURL(`${baseURL}/templates/invalid-template`);
-    });
-
-    test('when user submits page with no data, then an error is displayed', async ({
-      page,
-    }) => {
-      const errorMessage = 'Select an option';
-
-      const previewLetterTemplatePage = new TemplateMgmtPreviewLetterPage(page);
-
-      await previewLetterTemplatePage.loadPage(templates.valid.id);
-
-      await previewLetterTemplatePage.clickContinueButton();
-
-      await expect(previewLetterTemplatePage.errorSummary).toBeVisible();
-
-      const selectOptionErrorLink =
-        previewLetterTemplatePage.errorSummary.locator(
-          '[href="#previewLetterTemplateAction"]'
-        );
-
-      await expect(selectOptionErrorLink).toHaveText(errorMessage);
-
-      await selectOptionErrorLink.click();
-
-      await expect(
-        page.locator('#previewLetterTemplateAction')
-      ).toBeInViewport();
     });
   });
 });
